+++ NEWS.md
@@ -1,5 +1,6 @@
 # GOSemSim 2.39.2
 
++ fix `termSim(..., method = "TCSS")` for terms belonging to more than one cluster, where `ica[[clusid[[term]]]]` attempted recursive indexing ("subscript out of bounds") (2026-09-21, Mon)
 + add `set_auto_update()` to control auto-update of ontology databases (2026-06-29, Mon)
 
 # GOSemSim 2.39.1